image-proccessing/pyproject.toml
[tool.poetry]
name = "image-proccessing"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
opencv-python = "^4.0"
grpcio = "^1.19"
numpy = "^1.16"
grpcio-tools = "^1.19"
python-dotenv = "^0.10.1"
grpcio-reflection = "^1.19"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
mypy = "^0.670.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
